fix differences in behavior due to the loading of `request`
JSON.parse("") and JSON.parse("A string") now fail but were not
before. Moreover, `request` rejects colons (`:`) in the query string
while rest-client was passing them through.

It adds an URI encoding step (which we should have done in
the first place). However,  `encodeURI` does NOT encode `:`
so do it explicitely