use the autoproj configuration(s) as guards for the workspace packages
This should improve the interaction between an autoproj workspace
and additional folders and/or other workspaces. The folder packages
are now added *after* the autoproj build configuration, and *before*
a possible follow-up package that would not be part of the workspace.
It ensures that the packages of a given workspace are grouped
together, and that two open workspaces would not be interleaved.