forcefully kill subprocesses on dispose
There's no way to do it through the normal task-terminate mechanism,
since when vscode shuts down an extension host, it does give
us a way to wait on promises.

We still attempt to terminate for the "normal" case, that is the
case where all folders from a workspace have been removed.